fix: moved util functions to util.js
