+++ frontend/src/utils/index.js
@@ -247,7 +247,7 @@ export function _eval(code, context = {}) {
   }
 }
 
-export function evaluate_depends_on_value(expression, doc) {
+export function evaluateDependsOnValue(expression, doc) {
   if (!expression) return true
   if (!doc) return true
 
@@ -273,4 +273,21 @@ export function evaluate_depends_on_value(expression, doc) {
   }
 
   return out
+}
+
+export function convertSize(size) {
+  const units = ['B', 'KB', 'MB', 'GB', 'TB']
+  let unitIndex = 0
+  while (size > 1024) {
+    size /= 1024
+    unitIndex++
+  }
+  return `${size?.toFixed(2)} ${units[unitIndex]}`
+}
+
+export function isImage(extention) {
+  if (!extention) return false
+  return ['png', 'jpg', 'jpeg', 'gif', 'svg', 'bmp', 'webp'].includes(
+    extention.toLowerCase(),
+  )
 }
